refactor: ObjectMapper가 RowMapper를 상속받는 형태로 수정
